Sumedh said to make changes

+++ aruw-mcb-project/src/aruwsrc/robot/engineer/cube_down_command.hpp
@@ -30,7 +30,7 @@ class CubeDownCommand : public tap::control::Command
 {
 public: 
     CubeDownCommand(CubeStorageSubsystem &cubeLift);
-    int32_t power = 1;
+    int16_t power = 1;
 
 
     void initialize() override;
